Adding some new git aliases, updating the LunarVim installation

+++ submodules.sh
@@ -5,4 +5,5 @@
 git submodule init
 git submodule update
 
-ln -s $PWD/LunarVim ~/.config/nvim
+[ ! -d "~/.config" ] && mkdir -p ~/.config
+[[ ! -L "~/.config/nvim" && ! -d "~/.config/nvim" ]] && ln -s $PWD/LunarVim ~/.config/nvim
